컴포넌트 분리 및 라우터 적용
